style(web): vertically align text with images in links
Adjust the CSS in index_stu_left.jsp to ensure that the text within links is
vertically centered with the accompanying images. This improves the visual
consistency of the page by aligning the text and icon elements.Additionally, introduce line-height property to the paragraph tag for better
control over the spacing between text and icons.